+++ crates/parse_selectors/src/regexes/mod.rs
@@ -184,6 +184,7 @@ lazy_static! {
 				| \.getElementsByClassName
 				| \.classList\s*+\.(?> add | remove | contains | replace | toggle )
 				| \.setAttribute
+				| window\s*+\.(?> location\.assign | location\.replace | open )
 			)
 			(?<join>
 				\(\s*+