+++ crates/parse_selectors/tests/js/window-open/output.js
@@ -1,0 +1,34 @@
+window.open("#a")
+window.open("#a");
+window.open("/#a");
+window.open('#b')
+window.open('#b');
+window.open('/#b');
+window.open('#c')
+window.open('#c');
+window.open('/#c');
+
+window.open( '#d' )
+window.open( '#d' );
+window.open( '/#d' );
+window.open(  '#d'  )
+window.open(  '#d'  );
+window.open(  '/#d'  );
+
+window
+	.open('#e')
+window
+	.open('#e');
+window
+	.open('/#e');
+window
+	.open(  '#f')
+window
+	.open(  '#f');
+window
+	.open(  '/#f');
+
+window.open('/path/#g')
+window.open('path/#g')
+window.open('https://www.rust-lang.org/#foo')
+window.open('http://www.rust-lang.org/#foo')