+++ crates/parse_selectors/tests/js/window-open/source.js
@@ -1,0 +1,34 @@
+window.open("#SELECTOR-1")
+window.open("#SELECTOR-1");
+window.open("/#SELECTOR-1");
+window.open('#SELECTOR-2')
+window.open('#SELECTOR-2');
+window.open('/#SELECTOR-2');
+window.open('#SELECTOR-3')
+window.open('#SELECTOR-3');
+window.open('/#SELECTOR-3');
+
+window.open( '#SELECTOR-4' )
+window.open( '#SELECTOR-4' );
+window.open( '/#SELECTOR-4' );
+window.open(  '#SELECTOR-4'  )
+window.open(  '#SELECTOR-4'  );
+window.open(  '/#SELECTOR-4'  );
+
+window
+	.open('#SELECTOR-5')
+window
+	.open('#SELECTOR-5');
+window
+	.open('/#SELECTOR-5');
+window
+	.open(  '#SELECTOR-6')
+window
+	.open(  '#SELECTOR-6');
+window
+	.open(  '/#SELECTOR-6');
+
+window.open('/path/#SELECTOR-7')
+window.open('path/#SELECTOR-7')
+window.open('https://www.rust-lang.org/#foo')
+window.open('http://www.rust-lang.org/#foo')