+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Updates
 - Parse and encode minify-selectors specific prefixed selectors [#11](https://github.com/adamgian/minify-selectors/issues/11)
 - Add support for JS Location interface proprties: `window.location.hash`, 'window.location.href' and `window.location`
+- Add support for JS URL methods: `window.open()`
 
 ### Fixes
 - Only encode target IDs for relative URLs (i.e. without the protocol)